ecs-ec2-private/requirements.txt
attrs==19.3.0
aws-cdk.assets==1.35.0
aws-cdk.aws-apigateway==1.35.0
aws-cdk.aws-applicationautoscaling==1.35.0
aws-cdk.aws-autoscaling==1.35.0
aws-cdk.aws-autoscaling-common==1.35.0
aws-cdk.aws-autoscaling-hooktargets==1.35.0
aws-cdk.aws-batch==1.35.0
aws-cdk.aws-certificatemanager==1.35.0
aws-cdk.aws-cloudformation==1.35.0
aws-cdk.aws-cloudfront==1.35.0
aws-cdk.aws-cloudwatch==1.35.0
aws-cdk.aws-codebuild==1.35.0
aws-cdk.aws-codecommit==1.35.0
aws-cdk.aws-codedeploy==1.35.0
aws-cdk.aws-codepipeline==1.35.0
aws-cdk.aws-codepipeline-actions==1.35.0
aws-cdk.aws-ec2==1.35.0
aws-cdk.aws-ecr==1.35.0
aws-cdk.aws-ecr-assets==1.35.0
aws-cdk.aws-ecs==1.35.0
aws-cdk.aws-efs==1.35.0
aws-cdk.aws-elasticloadbalancing==1.35.0
aws-cdk.aws-elasticloadbalancingv2==1.35.0
aws-cdk.aws-events==1.35.0
aws-cdk.aws-events-targets==1.35.0
aws-cdk.aws-iam==1.35.0
aws-cdk.aws-kms==1.35.0
aws-cdk.aws-lambda==1.35.0
aws-cdk.aws-logs==1.35.0
aws-cdk.aws-route53==1.35.0
aws-cdk.aws-route53-targets==1.35.0
aws-cdk.aws-s3==1.35.0
aws-cdk.aws-s3-assets==1.35.0
aws-cdk.aws-sam==1.35.0
aws-cdk.aws-secretsmanager==1.35.0
aws-cdk.aws-servicediscovery==1.35.0
aws-cdk.aws-sns==1.35.0
aws-cdk.aws-sns-subscriptions==1.35.0
aws-cdk.aws-sqs==1.35.0
aws-cdk.aws-rds==1.35.0
aws-cdk.aws-ssm==1.35.0
aws-cdk.aws-stepfunctions==1.35.0
aws-cdk.core==1.35.0
aws-cdk.cx-api==1.35.0
aws-cdk.region-info==1.35.0
cattrs==1.0.0
constructs==3.0.2
jsii==1.4.1
publication==0.0.3
python-dateutil==2.8.1
six==1.14.0
typing-extensions==3.7.4.2
